+++ sprockets/http/runner.py
@@ -50,11 +50,13 @@ class Runner(object):
         except AttributeError:
             setattr(self.application, 'runner_callbacks', {'shutdown': []})
 
-    def start_server(self, port_number):
+    def start_server(self, port_number, number_of_procs=0):
         """
         Create a HTTP server and start it.
 
         :param int port_number: the port number to bind the server to
+        :param int number_of_procs: number of processes to pass to
+            Tornado's ``httpserver.HTTPServer.start``.
 
         If the application's ``debug`` setting is ``True``, then we are
         going to run in a single-process mode; otherwise, we'll let
@@ -73,20 +75,22 @@ class Runner(object):
                 'log_function', sprockets.logging.tornado_log_function)
             self.logger.info('starting processes on port %d', port_number)
             self.server.bind(port_number)
-            self.server.start(0)
+            self.server.start(number_of_procs)
 
-    def run(self, port_number):
+    def run(self, port_number, number_of_procs=0):
         """
         Create the server and run the IOLoop.
 
         :param int port_number: the port number to bind the server to
+        :param int number_of_procs: number of processes to pass to
+            Tornado's ``httpserver.HTTPServer.start``.
 
         If the application's ``debug`` setting is ``True``, then we are
         going to run in a single-process mode; otherwise, we'll let
         tornado decide how many sub-processes to spawn.
 
         """
-        self.start_server(port_number)
+        self.start_server(port_number, number_of_procs)
         ioloop.IOLoop.instance().start()
 
     def _on_signal(self, signo, frame):